refactor(tmdb): introduce new catalog system for providers
Refactored code, focusing on the TMDB model, to introduce a new catalog system for custom providers. This allows developers and users to configure and dynamically choose which films appear in the app.

- Reconfigured provider system to support new changes/features and have their own catalogs
- TMDB remains the main source of metadata/films provider
- Added initial unit test for TMDB repository
- This commit is depressing af

This refactoring lays the groundwork for more flexible content management and will be followed by additional unit tests to reduce build times and ensure code quality.


Former-commit-id: 2ecd1dd09d284547b478bcd8a3930ccb0c0df197