fix(provider): correct provider toggle and re-trigger catalogs
Fixed a bug where enabling/disabling a provider did not update the API map of the player, causing the player to still use it as a link scraper rather than the next available one.

Also fixed the catalog feature to ensure that enabling or disabling a provider correctly re-triggers updates on both the search and home screens, allowing new catalogs to be displayed.


Former-commit-id: 0c98b75d69a4b4ea2288b8048a35a2e424b67dff